fixed routing issues

+++ products/models.py
@@ -5,7 +5,7 @@ from django.db import models
 
 class Painting(models.Model):
     name = models.CharField(max_length=254)
-    year = models.DecimalField(max_digits=20, decimal_places=2)
+    year = models.IntegerField()
     price = models.DecimalField(max_digits=20, decimal_places=2)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
